Move ``_AutodocAttrGetter`` to ``_shared.py``
