+++ sphinx/ext/autodoc/_docstrings.py
@@ -25,8 +25,8 @@ if TYPE_CHECKING:
 
     from sphinx.events import EventManager
     from sphinx.ext.autodoc._directive_options import _AutoDocumenterOptions
-    from sphinx.ext.autodoc._importer import _AttrGetter
     from sphinx.ext.autodoc._property_types import _ItemProperties
+    from sphinx.ext.autodoc._shared import _AttrGetter
 
 logger = logging.getLogger('sphinx.ext.autodoc')
