+++ sphinx/ext/autodoc/_loader.py
@@ -46,8 +46,9 @@ if TYPE_CHECKING:
     from sphinx.environment import _CurrentDocument
     from sphinx.events import EventManager
     from sphinx.ext.autodoc._directive_options import _AutoDocumenterOptions
-    from sphinx.ext.autodoc._importer import _AttrGetter, _ImportedObject
+    from sphinx.ext.autodoc._importer import _ImportedObject
     from sphinx.ext.autodoc._property_types import _AutodocFuncProperty, _AutodocObjType
+    from sphinx.ext.autodoc._shared import _AttrGetter
 
 logger = logging.getLogger(__name__)
