+++ tests/test_ext_autodoc/test_ext_autodoc.py
@@ -26,7 +26,7 @@ from sphinx.ext.autodoc._generate import _generate_directives
 from sphinx.ext.autodoc._loader import _load_object_by_name
 from sphinx.ext.autodoc._property_types import _ItemProperties
 from sphinx.ext.autodoc._sentinels import ALL
-from sphinx.ext.autodoc.directive import _AutodocAttrGetter
+from sphinx.ext.autodoc._shared import _AutodocAttrGetter
 from sphinx.util.inspect import safe_getattr
 
 from tests.test_ext_autodoc.autodoc_util import do_autodoc
